refactor: import two-factor and passkey api through their domain index
