+++ frontend/src/pages/auth/components/SignupFactorSetup.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { motion } from 'motion/react';
-import type { StepUpPayload } from '@/api/two-factor/types';
+import type { StepUpPayload } from '@/api/two-factor';
 import { PasskeyEnrollment } from '@/components/passkeys/PasskeyEnrollment';
 import { TotpEnrollment } from '@/components/two-factor/TotpEnrollment';
 import { TwoFactorModalShell } from '@/components/two-factor/TwoFactorModalShell';
